@@ -6,75 +6,81 @@
 	"type": "module",
 	"private": true,
 	"scripts": {
-		"build": "tsc && node bootstrap.cjs vite build",
+		"build": "tsc && tsx bootstrap vite build",
 		"dev": "yarn start",
 		"lint": "tsc",
-		"preview": "tsc && node bootstrap.cjs vite preview",
-		"start": "node bootstrap.cjs vite",
-		"test": "node bootstrap.cjs vitest run",
-		"test:e2e": "node bootstrap.cjs yarn playwright test",
+		"preview": "tsc && tsx bootstrap vite preview",
+		"start": "tsx bootstrap vite",
+		"test": "tsx bootstrap vitest run",
+		"test:bench": "tsx bootstrap vitest bench",
+		"test:coverage": "tsx bootstrap vitest run --coverage",
 		"test:e2e:setup": "yarn playwright install --with-deps",
-		"test:bench": "node bootstrap.cjs vitest bench",
-		"test:coverage": "node bootstrap.cjs vitest run --coverage",
-		"test:watch": "node bootstrap.cjs vitest watch --ui"
+		"test:e2e": "tsx bootstrap yarn playwright test",
+		"test:watch": "tsx bootstrap vitest watch --ui",
+		"update": "yarn upgrade-interactive"
 	},
 	"dependencies": {
-		"@dnd-kit/core": "^6.1.0",
-		"@dnd-kit/sortable": "^8.0.0",
-		"@dnd-kit/utilities": "^3.2.2",
-		"@linaria/core": "6.1.0",
-		"@linaria/react": "6.1.0",
-		"@tanstack/react-router": "^1.19.9",
-		"@tanstack/react-store": "^0.3.1",
-		"@tanstack/router-devtools": "^1.19.9",
-		"@tanstack/router-vite-plugin": "^1.19.6",
-		"dayjs": "^1.11.10",
-		"lodash-es": "^4.17.21",
-		"plausible-tracker": "^0.3.8",
-		"react": "^18.2.0",
-		"react-device-detect": "^2.2.3",
-		"react-dom": "^18.2.0",
-		"sass": "^1.71.1"
+		"@dnd-kit/core": "6.1.0",
+		"@dnd-kit/sortable": "8.0.0",
+		"@dnd-kit/utilities": "3.2.2",
+		"@linaria/atomic": "6.2.0",
+		"@linaria/core": "6.2.0",
+		"@linaria/react": "6.2.1",
+		"@tanstack/react-router": "1.58.15",
+		"@tanstack/react-store": "0.5.5",
+		"@tanstack/router-devtools": "1.58.15",
+		"@tanstack/router-plugin": "1.58.12",
+		"@use-gesture/react": "10.3.1",
+		"dayjs": "1.11.13",
+		"lodash-es": "4.17.21",
+		"plausible-tracker": "0.3.9",
+		"react": "19.0.0-rc-67fee58b-20240926",
+		"react-device-detect": "2.2.3",
+		"react-dom": "19.0.0-rc-67fee58b-20240926",
+		"sass": "1.79.4"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.24.0",
-		"@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-		"@playwright/test": "^1.42.1",
-		"@rollup/pluginutils": "^5.1.0",
-		"@testing-library/jest-dom": "^6.4.2",
-		"@testing-library/react": "^14.2.1",
-		"@testing-library/react-hooks": "^8.0.1",
-		"@testing-library/user-event": "^14.5.2",
-		"@trivago/prettier-plugin-sort-imports": "^4.3.0",
-		"@types/css": "^0.0.37",
-		"@types/lodash-es": "^4.17.12",
-		"@types/node": "^20.11.7",
-		"@types/react": "^18.2.61",
-		"@types/react-dom": "^18.2.19",
-		"@vitejs/plugin-react": "^4.2.1",
-		"@vitest/coverage-v8": "^1.3.1",
-		"@vitest/ui": "^1.3.1",
-		"@wyw-in-js/babel-preset": "0.5.0",
-		"@wyw-in-js/shared": "0.5.0",
-		"@wyw-in-js/transform": "0.5.0",
-		"@wyw-in-js/vite": "0.5.0",
-		"babel-preset-react-app": "^10.0.1",
-		"cross-env": "^7.0.3",
-		"css": "^3.0.0",
-		"happy-dom": "13.7.7",
-		"prettier": "3.2.5",
-		"rollup-plugin-workbox": "^8.1.0",
-		"typescript": "^5.4.2",
-		"vite": "^5.1.6",
-		"vite-bundle-visualizer": "^1.0.1",
-		"vite-tsconfig-paths": "^4.3.1",
-		"vitest": "^1.3.1",
-		"workbox-build": "^7.0.0",
-		"workbox-core": "^7.0.0",
-		"workbox-expiration": "^7.0.0",
-		"workbox-precaching": "^7.0.0",
-		"workbox-routing": "^7.0.0",
-		"workbox-strategies": "^7.0.0"
+		"@babel/core": "7.25.2",
+		"@babel/plugin-proposal-private-property-in-object": "7.21.11",
+		"@playwright/test": "1.47.2",
+		"@rollup/pluginutils": "5.1.2",
+		"@testing-library/dom": "10.4.0",
+		"@testing-library/jest-dom": "6.5.0",
+		"@testing-library/react": "16.0.1",
+		"@testing-library/react-hooks": "8.0.1",
+		"@testing-library/user-event": "14.5.2",
+		"@trivago/prettier-plugin-sort-imports": "4.3.0",
+		"@types/css": "0.0.38",
+		"@types/lodash-es": "4.17.12",
+		"@types/node": "20.16.10",
+		"@types/react": "18.3.10",
+		"@types/react-dom": "18.3.0",
+		"@vitejs/plugin-react": "4.3.2",
+		"@vitest/coverage-v8": "2.1.1",
+		"@vitest/ui": "2.1.1",
+		"@wyw-in-js/babel-preset": "0.5.4",
+		"@wyw-in-js/shared": "0.5.4",
+		"@wyw-in-js/transform": "0.5.4",
+		"@wyw-in-js/vite": "0.5.4",
+		"babel-plugin-react-compiler": "0.0.0-experimental-57db2f6-20240926",
+		"babel-preset-react-app": "10.0.1",
+		"cross-env": "7.0.3",
+		"css": "3.0.0",
+		"happy-dom": "15.7.4",
+		"prettier": "3.3.3",
+		"rollup-plugin-workbox": "8.1.0",
+		"tsx": "4.19.1",
+		"typescript": "5.6.2",
+		"vite": "5.4.8",
+		"vite-bundle-visualizer": "1.2.1",
+		"vite-tsconfig-paths": "5.0.1",
+		"vitest": "2.1.1",
+		"workbox-build": "7.1.1",
+		"workbox-core": "7.1.0",
+		"workbox-expiration": "7.1.0",
+		"workbox-precaching": "7.1.0",
+		"workbox-routing": "7.1.0",
+		"workbox-strategies": "7.1.0"
 	},
 	"browserslist": {
 		"production": [
@@ -91,7 +97,7 @@
 		"url": "https://github.com/hmerritt/adrift.git"
 	},
 	"engines": {
-		"node": ">=18.0.0"
+		"node": ">=20.0.0"
 	},
-	"packageManager": "yarn@3.5.0"
+	"packageManager": "yarn@4.5.0"
 }
